+++ src/sn/Areas.ts
@@ -8,24 +8,19 @@
 import {IAreas} from './CmnInterface';
 
 export class Areas implements IAreas {
-	hAreas	: {[name: number]: number}	= {};
+	hAreas	: {[name: string]: number}	= Object.create(null);
 
 	clear() {this.hAreas = {};}
 
 	search(idx: number): boolean {
-		for (const begin in this.hAreas) {
-			if (idx < parseInt(begin)) continue;
-
-			if (idx <= this.hAreas[begin]) return true;
-		}
-
-		return false;
+		return Object.entries(this.hAreas)
+		.some(([begin, v])=> (idx >= parseInt(begin) && idx <= v));
 	}
 
 	record(idx: number): void {
 		if (this.search(idx)) return;
 
-		for (const begin in this.hAreas) {
+		for (const begin of Object.keys(this.hAreas)) {
 			// 領域末尾+1とマッチ
 			if (this.hAreas[begin] +1 === idx) {
 				if ((idx +1) in this.hAreas) {
@@ -60,9 +55,8 @@ export class Areas implements IAreas {
 			return;
 		}
 
-		for (const begin in this.hAreas) {
-			if (idx < parseInt(begin)) continue;
-			if (this.hAreas[begin] < idx) continue;
+		for (const [begin, v] of Object.entries(this.hAreas)) {
+			if (idx < parseInt(begin) || v < idx) continue;
 
 			// 領域末尾とマッチ
 			if (this.hAreas[begin] === idx) {
@@ -71,7 +65,7 @@ export class Areas implements IAreas {
 			}
 
 			// 領域内とマッチ、二つの領域に分割
-			this.hAreas[idx +1] = this.hAreas[begin];
+			this.hAreas[idx +1] = v;
 			this.hAreas[begin] = idx -1;
 			return;
 		}
@@ -80,21 +74,17 @@ export class Areas implements IAreas {
 	get count(): number {return Object.keys(this.hAreas).length;}	// 4tst
 
 	toString(): string {
-		let ret = '';
+		let ret = ',';
 
-		const aBegin: number[] = [];
-		for (const begin in this.hAreas) aBegin.push(parseInt(begin));
-		aBegin.sort(function (x: number, y: number): number {return x-y;});
+		for (const k of Object.keys(this.hAreas)
+		.map(k=> parseInt(k))
+		.sort((x, y)=> x-y)) {
+			ret += k === this.hAreas[k]
+			? ','+ k
+			: ','+ k +'~'+ this.hAreas[k];
+		}
 
-		aBegin.forEach(v=> {ret += ','+ v
-			+ (v === this.hAreas[v])
-				? ''
-				: '~'+ this.hAreas[v];
-		});
-
-		if (ret !== '') ret = ret.slice(1);
-
-		return ret;
+		return ret.slice(1);
 	}
 
 }