+++ src/sn/SysWeb.ts
@@ -87,9 +87,11 @@ export class SysWeb extends SysBase {
 
 		const src = await res.text();
 		const oJs = JSON.parse(this.decStr(fn, src));
-		for (const nm in oJs) {
-			const h = hPathFn2Exts[nm] = oJs[nm];
-			for (const ext in h) if (ext !== ':cnt') h[ext] = this.arg.cur + h[ext];
+		for (const [nm, v] of Object.entries(oJs)) {
+			const h = hPathFn2Exts[nm] = <any>v;
+			for (const [ext, w] of Object.entries(h)) {
+				if (ext !== ':cnt') h[ext] = this.arg.cur + w;
+			}
 		}
 	}